requirements.txt
alabaster==0.7.12
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.13.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.0.4
crispy-bootstrap5==0.7
cryptography==39.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.0.8
django-allauth==0.52.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-coverage-plugin==3.0.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.8.0
docutils==0.19
drf-spectacular==0.25.1
exceptiongroup==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.4
filelock==3.9.0
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
hiredis==2.1.0
httptools==0.5.0
identify==2.5.12
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.11
ipython==8.8.0
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.9.0
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
phonenumbers==8.13.6
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
termcolor==2.2.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.0
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing-extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.17.1
watchdog==2.2.1
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
